% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_analysis.R
\name{parse_log_filename}
\alias{parse_log_filename}
\title{Parse ShinyProxy Log File Name}
\usage{
parse_log_filename(filename)
}
\arguments{
\item{filename}{Character. The full path to the log file.}
}
\value{
A data frame with the columns \code{specId}, \code{proxyId}, \code{startupTime}, and \code{logType}.
}
\description{
This function parses the name of a log file generated by ShinyProxy and extracts metadata:
\code{specId}, \code{proxyId}, \code{startupTime}, and \code{logType}.
}
\examples{
\dontrun{
  parse_log_filename(
  "path/to/containersLogs/log_file_(stdout|stderr).log"
  )
}
}
