% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_analysis.R
\name{parse_log_filename}
\alias{parse_log_filename}
\title{Parse ShinyProxy Log File Name}
\usage{
parse_log_filename(filename)
}
\arguments{
\item{filename}{Character. The full path to the log file.}
}
\value{
A data frame with the columns \code{specId}, \code{proxyId}, \code{startupTime}, and \code{logType}.
}
\description{
This function parses the name of a log file generated by ShinyProxy and extracts metadata:
\code{specId}, \code{proxyId}, \code{startupTime}, and \code{logType}.
}
\examples{
parse_log_filename("../../shinyproxy/shinyproxy/container-logs/logs/crash_simulator_43c8bf13-051f-4fe9-8362-a6b83016a218_05_Feb_2025_18_30_53_stderr.log")
}
